requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
sqlalchemy==2.0.35
openai==1.45.0
python-dotenv==1.0.1
